change(**): Move query/fragment to .ts file
