+++ src/hooks/VirtualizedList.hooks.ts
@@ -1,10 +1,38 @@
+import { gql } from "@apollo/client";
 import { useMemo, useState } from "react";
 import {
-    GetPostConnectionDocument,
-    GetPostConnectionQueryVariables,
-    PostFragment as Post,
-    useGetPostConnectionQuery
-} from "./__generated__/post-connection-query.generated";
+  GetPostConnectionDocument,
+  GetPostConnectionQueryVariables,
+  PostFragment as Post,
+  useGetPostConnectionQuery
+} from "./__generated__/VirtualizedList.hooks.generated";
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+const _GET_POST_CONNECTION_QUERY = gql`
+  query GetPostConnection($first: Int, $after: String, $query: String) {
+    postConnection(first: $first, after: $after, query: $query) {
+      pageInfo {
+        hasNextPage
+        hasPreviousPage
+        startCursor
+        endCursor
+      }
+      totalCount
+      edges {
+        node {
+          ...Post
+        }
+        cursor
+      }
+    }
+  }
+  fragment Post on Post {
+    id
+    userId
+    title
+    body
+  }
+`;
 
 const DEFAULT_FETCH_SIZE = 10;
 const FIRST_CURSOR = "0";
