--- src/hooks/post-connection-query.graphql
@@ -1,23 +1,0 @@
-query GetPostConnection($first: Int, $after: String, $query: String) {
-  postConnection(first: $first, after: $after, query: $query) {
-    pageInfo {
-      hasNextPage
-      hasPreviousPage
-      startCursor
-      endCursor
-    }
-    totalCount
-    edges {
-      node {
-        ...Post
-      }
-      cursor
-    }
-  }
-}
-fragment Post on Post {
-  id
-  userId
-  title
-  body
-}
